+++ Adj.hs
@@ -1,4 +1,5 @@
 module Adj (module Exports) where
 
+import Adj.Auxiliary as Exports
 import Adj.Algebra as Exports
 import Adj.Program as Exports