+++ CHANGELOG.md
@@ -61,3 +61,5 @@
 * Define `-|-|` operator in `Functor` typeclass
 * Define `-|-|-|` operator in `Functor` typeclass
 * Rename `Disbandable` typeclass and module to `Casting`
+* Define `Adj.Auxiliary` module
+* Move `Casting` typeclass to `Adj.Auxiliary` method